requirements.txt
aiohttp==3.7.3
alpha-vantage==2.2.0
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-environ==0.4.5
django-heroku==0.3.1
environ==1.0
finnhub-python==2.3.0
googlefinance==0.7
gunicorn==20.0.4
idna==2.10
lxml==4.6.2
multidict==5.0.2
nasdaq-stock==0.0.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stockquotes==2.0.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
whitenoise==5.2.0
yarl==1.6.3
